extract `AutomaticReportingTable` to component
